+++ server/app.ts
@@ -1,0 +1,18 @@
+import cors from 'cors';
+import express from 'express';
+import availabilityRouter from './routes/availability.js';
+import bookingRouter from './routes/booking.js';
+import bookingsRouter from './routes/bookings.js';
+import healthRouter from './routes/health.js';
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+
+app.use('/healthz', healthRouter);
+app.use('/api/availability', availabilityRouter);
+app.use('/api/bookings', bookingsRouter);
+app.use('/api/book', bookingRouter);
+
+export default app;
